% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_plate_results.R
\name{create_plate_results4}
\alias{create_plate_results4}
\title{Creating plate results from FOV filtered lists}
\usage{
create_plate_results4(
  data_filtered_list,
  feature_for_filter = feature_for_filter,
  objective = c("objective_40x", "objective_63x", "objective_20x", "objective_10x"),
  number_of_objects = NULL,
  Mean_per_Well = NULL,
  StdDev_per_Well = NULL,
  Sum_per_Well = NULL,
  Max_per_Well = NULL,
  Min_per_Well = NULL,
  Median_per_Well = NULL,
  spots_per_length_neurite = NULL,
  Neurite_Length_per_neuron = NULL,
  xx_segments_per_neurons = NULL,
  proportion_x_y = NULL,
  ...
)
}
\arguments{
\item{data_filtered_list}{list of data containing measurements per field of view, filtered for a feature}

\item{feature_for_filter}{name of an element in data_list for which data should be filtered, should be the same as used in filter_data_by_feature}

\item{objective}{name of the objective used for the analysis ("objective_40x", "objective_63x")}

\item{number_of_objects}{character vector of features that should be counted}

\item{Mean_per_Well}{character vector containing names or partial names of the columns of which the mean should be calculated}

\item{StdDev_per_Well}{character vector containing names or partial names of the columns of which the standard deviation should be calculated}

\item{Sum_per_Well}{character vector containing names or partial names of the columns of which the sum should be calculated}

\item{Max_per_Well}{character vector containing names or partial names of the columns of which the max should be calculated}

\item{Min_per_Well}{character vector containing names or partial names of the columns of which the min should be calculated}

\item{Median_per_Well}{character vector containing names or partial names of the columns of which the median should be calculated}

\item{spots_per_length_neurite}{character vector containing names or partial names of feature/s for which the number of spots per neurite length should be calculated (per 100 µm)}

\item{Neurite_Length_per_neuron}{TRUE or FALSE, should the neurite length per neuron be calculated. Default is NULL.}

\item{xx_segments_per_neurons}{character vector containing names or partial names of feature/s for which the sum of segment length/number of neuronal nuclei should be calculated}

\item{proportion_x_y}{character vector of length 2, or list of character vectors of length 2; first element of the vector will be devided by the second element of the vector}

\item{...}{additional arguments that can be passed to the function in order to add them to plate results. E.g. Cell_line, Compound, Replicate; if passed as data.frame they must contain the Columns Row and Column}
}
\value{
data frame with plate results per well
}
\description{
Creating plate results from FOV filtered lists
}
