% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_data_by_feature.R
\name{filter_data_by_feature}
\alias{filter_data_by_feature}
\title{Filter data per field of view by the number of nuclei present per field of view}
\usage{
filter_data_by_feature(
  data_list,
  feature_for_filter = "data_neuronal_nuclei",
  min = 0,
  max = 200
)
}
\arguments{
\item{data_list}{list of data containing measurements per field of view}

\item{feature_for_filter}{name of an element in data_list for which data should be filtered, default is "data_neuronal_nuclei"}

\item{min}{minimum number of nuclei to keep field of view. default = 0}

\item{max}{maximum number of nuclei to keep field of view. default = 200}
}
\value{
List of filtered data for each of the five measurements
}
\description{
Filter data per field of view by the number of nuclei present per field of view
}
