% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_data_by_number_of_nuclei.R
\name{filter_data_by_number_of_nuclei}
\alias{filter_data_by_number_of_nuclei}
\title{Filter data per field of view by the number of nuclei present per field of view}
\usage{
filter_data_by_number_of_nuclei(data_list, min = 0, max = 200)
}
\arguments{
\item{data_list}{list of data containing five list elements (data_neuronal_nuclei, data_SV2_neurites, data_coloc_punctae, data_homer_spots, data_neurite_segemnts)}

\item{min}{minimum number of nuclei to keep field of view. default = 0}

\item{max}{maximum number of nuclei to keep field of view. default = 200}
}
\value{
List of filtered data for each of the five measurements
}
\description{
Filter data per field of view by the number of nuclei present per field of view
}
