% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering_report_Compounds.R
\name{filtering_report_Compounds}
\alias{filtering_report_Compounds}
\title{Filter data per field of view by the number of nuclei present per field of view}
\usage{
filtering_report_Compounds(
  data_list,
  feature_for_filter,
  data_filtered_list,
  Compound
)
}
\arguments{
\item{data_list}{list of data containing measurements per field of view}

\item{feature_for_filter}{name of an element in data_list for which data should be filtered, default is "data_neuronal_nuclei"}

\item{data_filtered_list}{name of the filtered data_list}

\item{Compound}{Compound list including columns "Row", "Column", "Compound"}
}
\value{
Statement, which Compounds were already missing in the input data and which compounds were removed in the filtering step
}
\description{
Filter data per field of view by the number of nuclei present per field of view
}
