% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_partial_columns.R
\name{find_partial_columns}
\alias{find_partial_columns}
\title{Find column names and respective list for arguments given to create_plate_results}
\usage{
find_partial_columns(data_filtered_list, target_names)
}
\arguments{
\item{data_filtered_list}{list of data containing measurements per field of view, filtered for a feature}

\item{target_names}{target names that should be column names or partial column names of one of the measurements data frames}
}
\value{
List with list name and column name for each target
}
\description{
Find column names and respective list for arguments given to create_plate_results
}
