% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_partial_list_elements.R
\name{find_partial_list_elements}
\alias{find_partial_list_elements}
\title{Find list element names for arguments given to create_plate_results to calculate number of objects}
\usage{
find_partial_list_elements(data_filtered_list, number_of_objects)
}
\arguments{
\item{data_filtered_list}{list of data containing measurements per field of view, filtered for a feature}

\item{target_names}{target names that should be column names or partial column names of one of the measurements data frames}
}
\value{
List with list name and column name for each target
}
\description{
Find list element names for arguments given to create_plate_results to calculate number of objects
}
