% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda_coeffs_sorted_all_features.R
\name{lda_coeffs_sorted_all_features}
\alias{lda_coeffs_sorted_all_features}
\title{Create z-prime plot per replicate including increasing numbers of features in the calculation}
\usage{
lda_coeffs_sorted_all_features(meta_control, features_control)
}
\arguments{
\item{meta_control}{dataframe with metadata (Replicate, cell_line etc), only control samples}

\item{features_control}{dataframe of scaled features per sample, only control samples}
}
\value{
data frame with features sorted by the absolute LD1 value
}
\description{
Create z-prime plot per replicate including increasing numbers of features in the calculation
}
