% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_distribution_before_after_filtering.R
\name{plots_distribution_before_after_filtering}
\alias{plots_distribution_before_after_filtering}
\title{Plot data points per field of view as histogram}
\usage{
plots_distribution_before_after_filtering(
  data_list,
  feature_for_filter,
  min = 0,
  max = 200
)
}
\arguments{
\item{data_list}{list of data containing measurements per field of view}

\item{feature_for_filter}{name of an element in data_list for which data should be filtered, usually neuronal nuclei}

\item{min}{minimum number of filtering feature (nuclei) to keep field of view. default = 0}

\item{max}{maximum number of filtering feature (nuclei) to keep field of view. default = 200}
}
\value{
Histograms showing the distribution of number of measurements per field of view
}
\description{
Plot data points per field of view as histogram
}
