% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_prime_elbow.R
\name{z_prime_elbow}
\alias{z_prime_elbow}
\title{Create z-prime plot per replicate including increasing numbers of features in the calculation}
\usage{
z_prime_elbow(features_scaled, meta)
}
\arguments{
\item{features_scaled}{dataframe of scaled features per sample}

\item{meta}{dataframe with metadata (Replicate, cell_line etc)}
}
\value{
Plot of LDA separation of WT and MUT per increased feature, plot of z-prime values per increased feature
}
\description{
Create z-prime plot per replicate including increasing numbers of features in the calculation
}
