% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_prime_per_replicate.R
\name{z_prime_per_replicate}
\alias{z_prime_per_replicate}
\title{Create z-prime value for each replicate}
\usage{
z_prime_per_replicate(LD1_values_df, zprime_per_replicate_df)
}
\arguments{
\item{LD1_df}{data.frame with LD1 value for each sample in a column called LD1, additionally cell_line, and Replicate columns have to be present}
}
\value{
vector of zprime values per replicate
}
\description{
Create z-prime value for each replicate
}
